프로필 완료(업로드동영상), 버그 수정

+++ src/controllers/videoController.js
@@ -1,6 +1,7 @@
+import User from "../models/User";
 import Video from "../models/Video";
 
-const isPromise = (v) => typeof v === "object" && typeof v.then === "function";
+// const isPromise = (v) => typeof v === "object" && typeof v.then === "function";
 
 export const home = async (req, res) => {
   // Video.find({}, (error, videos) => {
@@ -21,7 +22,7 @@ export const search = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
   if (!video)
     return res.status(404).render("404", { pageTitle: "Video not found" });
   return res.render("watch", { pageTitle: video.title, video });
@@ -30,17 +31,25 @@ export const watch = async (req, res) => {
 export const getEdit = async (req, res) => {
   const { id } = req.params;
   const video = await Video.findById(id);
-  if (!video)
+  if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== req.session.user._id) {
+    return res.status(403).redirect("/");
+  }
   return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
-  const video = await Video.exists({ id });
-  if (!video)
+  const video = await Video.findById({ id });
+  if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== req.session.user._id) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndUpdate(id, {
     title,
     description,
@@ -54,15 +63,25 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
-  const { title, description, hashtags } = req.body;
-  const { file } = req;
+  const {
+    session: {
+      user: { _id },
+    },
+    body: { title, description, hashtags },
+    file,
+  } = req;
+
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       fileUrl: file.path,
       title,
       description,
       hashtags: Video.formatHashtags(hashtags),
+      owner: _id,
     });
+    const owner = await User.findById(_id);
+    owner.videos.push(newVideo._id);
+    await owner.save();
     return res.redirect("/");
   } catch (error) {
     return res.status(400).render("upload", {
@@ -74,6 +93,13 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== req.session.user._id) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
